main/requirements.txt
cryptography>=41.0
Pillow>=10.0